backend/requirements.txt
fastapi
uvicorn[standard]
pandas
langchain-openai
langchain
langgraph
langchain-experimental
openpyxl
python-dotenv
boto3